New modification

+++ src/main/java/com/bibliotheque/services/LivreService.java
@@ -1,7 +1,9 @@
 package com.bibliotheque.services;
 
 import com.bibliotheque.models.Livre;
+import com.bibliotheque.models.Emprunt;
 import com.bibliotheque.repositories.LivreRepository;
+import com.bibliotheque.repositories.EmpruntRepository;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
 
@@ -14,6 +16,9 @@ public class LivreService {
     @Autowired
     private LivreRepository livreRepository;
 
+    @Autowired
+    private EmpruntRepository empruntRepository; 
+
     public List<Livre> getAllLivres() {
         return livreRepository.findAll();
     }
@@ -24,6 +29,12 @@ public class LivreService {
 
     public Livre saveLivre(Livre livre) {
         return livreRepository.save(livre);
+    }
+
+    public void rendreLivre(Long empruntId) {
+        Emprunt emprunt = empruntRepository.findById(empruntId).orElseThrow();
+        emprunt.setStatut("Terminé");
+        empruntRepository.save(emprunt);
     }
 
     public Livre updateLivre(Long id, Livre livreDetails) {
@@ -41,6 +52,10 @@ public class LivreService {
     }
 
     public void deleteLivre(Long id) {
+        // 🔹 Vérification si le livre est lié à un emprunt
+        if (empruntRepository.existsByLivreId(id)) {
+            throw new RuntimeException("❌ Ce livre ne peut pas être supprimé car il est lié à un emprunt !");
+        }
         livreRepository.deleteById(id);
     }
 }
